% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A string with the name of the file to be read}
}
\value{
If the file exists, the function returns a dataframe with the data from the .csv file
or an error message if it does not.
}
\description{
This function reads a datafile from the Fatality Anaysis Reporting System containing information
on fatal injuries from traffic accidents. The file is in CSV format.
}
