% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize data by month of the year.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector with years to be used when summarizing.}
}
\value{
A data frame for number of accidents per year and month.
}
\description{
This function groups and summarizes the data by month of the year
based on input regarding the years to be included.
}
