% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars.R
\name{make_filename}
\alias{make_filename}
\title{Generate file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer or string which indicates the year}
}
\description{
This function generates a name of the file using a particular year.
}
